[useTree]: more refactoring.
