+++ uui-core/src/data/processing/views/tree/constants.ts
@@ -13,3 +13,8 @@ export const PENDING_RECORD = Symbol('PENDING_RECORD');
 export const LOADING_RECORD = Symbol('LOADING_RECORD');
 export const LOADED_RECORD = Symbol('LOADED_RECORD');
 export const FAILED_RECORD = Symbol('FAILED_RECORD');
+
+export const PatchOrdering = {
+    TOP: Symbol('ORDERING_TOP'),
+    BOTTOM: Symbol('ORDERING_BOTTOM'),
+};
