+++ uui-core/src/data/processing/views/tree/index.ts
@@ -1,3 +1,2 @@
 export * from './hooks';
 export * from './exposed';
-export * from './PatchOrderingMap';
