+++ uui-core/src/data/processing/views/tree/treeState/types.ts
@@ -1,7 +1,7 @@
 import { CascadeSelection, DataSourceState, SortingOption } from '../../../../../types';
 import { LazyListViewProps } from '../../types';
 import { ItemsMap } from '../ItemsMap';
-import { TreeStructure } from '../treeStructure';
+import { TreeStructure } from '../treeStructure/TreeStructure';
 
 export interface LoadTreeOptions<TItem, TId, TFilter>
     extends Pick<LazyListViewProps<TItem, TId, TFilter>, 'api' | 'getChildCount' | 'filter' | 'fetchStrategy' | 'flattenSearchResults'> {
