+++ uui-core/src/data/processing/views/tree/types.ts
@@ -1,4 +1,4 @@
-import { FAILED_RECORD, LOADED_RECORD, LOADING_RECORD, NOT_FOUND_RECORD, PENDING_RECORD } from './constants';
+import { FAILED_RECORD, LOADED_RECORD, LOADING_RECORD, NOT_FOUND_RECORD, PENDING_RECORD, PatchOrdering } from './constants';
 
 export type RecordStatus =
 | typeof PENDING_RECORD
@@ -6,3 +6,7 @@ export type RecordStatus =
 | typeof LOADED_RECORD
 | typeof NOT_FOUND_RECORD
 | typeof FAILED_RECORD;
+
+export type PatchOrderingType =
+| typeof PatchOrdering.TOP
+| typeof PatchOrdering.BOTTOM;
